@@ -25,16 +25,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.1.0",
-    "@angular/cdk": "^18.1.0",
-    "@angular/common": "^18.1.0",
-    "@angular/compiler": "^18.1.0",
-    "@angular/core": "^18.1.0",
-    "@angular/forms": "^18.1.0",
-    "@angular/material": "^18.1.0",
-    "@angular/platform-browser": "^18.1.0",
-    "@angular/platform-browser-dynamic": "^18.1.0",
-    "@angular/router": "^18.1.0",
+    "@angular/animations": "^18.2.6",
+    "@angular/cdk": "^18.2.6",
+    "@angular/common": "^18.2.6",
+    "@angular/compiler": "^18.2.6",
+    "@angular/core": "^18.2.6",
+    "@angular/forms": "^18.2.6",
+    "@angular/material": "^18.2.6",
+    "@angular/platform-browser": "^18.2.6",
+    "@angular/platform-browser-dynamic": "^18.2.6",
+    "@angular/router": "^18.2.6",
     "@auth0/angular-jwt": "^5.2.0",
     "@cisstech/nge": "^17.7.1",
     "@myrmidon/auth-jwt-admin": "5.1.0",
@@ -68,13 +68,13 @@
     "rxjs": "~7.8.1",
     "ts-md5": "^1.3.1",
     "tslib": "^2.6.3",
-    "zone.js": "~0.14.7"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.1.0",
-    "@angular/cli": "^18.1.0",
-    "@angular/compiler-cli": "^18.1.0",
-    "@angular/localize": "^18.1.0",
+    "@angular-devkit/build-angular": "^18.2.6",
+    "@angular/cli": "^18.2.6",
+    "@angular/compiler-cli": "^18.2.6",
+    "@angular/localize": "^18.2.6",
     "@types/d3-dispatch": "^3.0.6",
     "@types/d3-drag": "^3.0.7",
     "@types/d3-timer": "^3.0.2",
